requirements.txt
numpy>=1.24,<1.25
scipy>=1.10,<1.11
matplotlib>=3.7,<3.8
pandas>=2.0,<2.1
pyyaml>=6.0
streamlit>=1.28,<1.39
plotly>=5.18,<6.0
